Cleaned up landing page scripts
