+++ public/scripts/landing-page/create-account.js
@@ -1,0 +1,64 @@
+export function handleAccountCreation() {
+    const [username, password] = getUsernamePasswordInput();
+
+    if (username.length === 0) {
+        setMessageText('Please provide a username.');
+    }
+    else if (password.length === 0) {
+        setMessageText('Please provide a password.');
+    }
+    else {
+        attemptAccountCreation(username, password);
+    }
+}
+
+/* TODO: Get rid of the status errors with in try catch
+    when function is completely implemented */
+async function attemptAccountCreation(username, password) {
+    try {
+        const response = await fetch('/api/createUser', {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+            body: JSON.stringify({
+                username: username,
+                password: password,
+            }),
+        });
+
+        if (response.status === 200) {
+            handleAccountCreationSuccess()
+        }
+        else if (response.status === 401) {
+            handleAccountCreationFail('Failed to create account. Status 400.');
+        }
+        else {
+            handleAccountCreationFail('Failed to create account. Please try again later.');
+        }
+
+    } catch (error) {
+        console.log('Error', error);
+        handleAccountCreationFail('Error creating account.');
+    }
+}
+
+function handleAccountCreationSuccess() {
+    setMessageText('User Account Created. Please Login.');
+}
+
+function handleAccountCreationFail(message) {
+    setMessageText(message);
+    usernameInput.value = '';
+    passwordInput.value = '';
+}
+
+function getUsernamePasswordInput() {
+    const usernameInput = document.getElementById('username-input');
+    const passwordInput = document.getElementById('password-input');
+    return [usernameInput.value, passwordInput.value]
+}
+
+function setMessageText(messageText) {
+    document.getElementById('message-text').innerHTML = messageText;
+}
